requirements.txt
astroid==2.3.3
autopep8==1.5.2
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
colorama==0.4.3
Flask==1.1.2
flask-marshmallow==0.12.0
Flask-SQLAlchemy==2.4.1
flask-swagger-ui==3.25.0
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
marshmallow==3.5.1
marshmallow-sqlalchemy==0.23.0
mccabe==0.6.1
psycopg2==2.8.5
pycodestyle==2.5.0
PyJWT==1.7.1
pylint==2.4.4
pylint-flask==0.6
pylint-flask-sqlalchemy==0.2.0
pylint-plugin-utils==0.6
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.16
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.11.2
